chore(RHIF-175): get rid of AppInfo with some code cleanup
